Fix abs() overload for mcu lms7002m calibrations
on ubuntu yakkety (gcc 6.2.0) abs(int8_t) had some resolution issue,
it was picking up a double type which it could not or with 0x40.
The fix was to force the integer version of abs with a cast.

+++ mcu_program/common_src/lms7002m_calibrations.c
@@ -65,7 +65,7 @@ static uint8_t toDCOffset(const int8_t offset)
 {
     if(offset >= 0)
         return offset;
-    return abs(offset) | 0x40;
+    return (uint8_t)(abs((int)offset) | 0x40);
 }
 
 static void FlipRisingEdge(const uint16_t addr, const uint8_t bits)
